AudioMiniPlayer: Update Media Info Collection View
This should fix these issues, wrong current info for the shuffled list and the media info cell size in the landscape mode.
Add functionality for infinite scrolling when RepeatAllItems mode is
active with handling also the previous action for infinite scrolling.
add limitation in dragging or declarating for only one cell per the
snap.