TabBarCoordinator: Enable the side bar navigation
The side bar navigation is now enabled on iPad running iPadOS 18
and versions above.

This is a wip prototype that was aimed to check the amount of work
needed in order to make this happen.

Currently each tab of the Audio section are made as proper navigation
tabs in that side bar only to avoid any huge refactor around the
PagingViewController and related classes.